incident_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-incident_env"
version = "0.1.0"
description = "IncidentEnv — OpenEnv environment for production incident response training"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "incident_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["incident_env", "incident_env.server"]
package-dir = { "incident_env" = ".", "incident_env.server" = "server" }
